+++ src/types.ts
@@ -84,10 +84,6 @@ export interface CombinableDiscountCounts {
 
 export type CountryCode = SupportedCountryCode | typeof REST_OF_WORLD;
 
-export enum DiscountApplicationStrategy {
-  First = 'FIRST',
-  Maximum = 'MAXIMUM',
-}
 export interface ProductOrCollectionResource extends Resource {
   title: string;
 }