+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # @shopify/discount-app-components
 
+## 2.1.2
+
+### Patch Changes
+
+- [#188](https://github.com/Shopify/discount-app-components/pull/188) [`0e366a8`](https://github.com/Shopify/discount-app-components/commit/0e366a8f9ede089de37ec4f8fcf0a86577c4711f) Thanks [@mathiusj](https://github.com/mathiusj)! - move strategy to constants
+
+## 2.1.1
+
+### Patch Changes
+
+- [#186](https://github.com/Shopify/discount-app-components/pull/186) [`4b47938`](https://github.com/Shopify/discount-app-components/commit/4b479380fb64cc4c730b9faff3baa22dc1227789) Thanks [@mathiusj](https://github.com/mathiusj)! - export strategy type from index
+
+## 2.1.0
+
+### Minor Changes
+
+- [#178](https://github.com/Shopify/discount-app-components/pull/178) [`0d67597`](https://github.com/Shopify/discount-app-components/commit/0d67597354622bd9fd616bfc7ca93dca1c013401) Thanks [@mathiusj](https://github.com/mathiusj)! - add strategy all to DiscountApplicationStrategyCard
+
 ## 2.0.0
 
 ### Major Changes